python-app/requirements.txt
# Core dependencies
streamlit>=1.28.0
anthropic>=0.7.0
snowflake-connector-python>=3.0.0
pandas>=1.5.0
python-dotenv>=1.0.0
cryptography>=3.4.8

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Additional utilities
typing-extensions>=4.0.0